% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_distribution.R
\name{add_point_coordinates}
\alias{add_point_coordinates}
\title{Add coordinates on horizontal plane}
\usage{
add_point_coordinates(spatial_data, reference_points)
}
\description{
Add coordinates of the ant clusters mapped onto the horizontal plane. Calculations
are performed based on the distances from the reference points to the cluster
center projected upward to the ground level. Because the exact geometrical solution is
often unavailable due to the distance measurement errors, the approximate solution
is achieved through the minimizing the loss function, which is proportional to the
differences between measured distances and those implied by the choice of the candidate center position.
}
\details{
@param spatial_data A data frame with the columns containing data on the distances
of the reference points to the cluster center shifted upward to the ground level.
@param reference_points A named list with the plane coordinates of the reference points.
List elements correspond to the different colonies.
}
