% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_analyses.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Interval for the proportion of ants}
\usage{
calculate_CI(data, alpha)
}
\arguments{
\item{data}{Data frame returned by \code{get_prob_mass}}

\item{alpha}{Numeric value setting the maximum value for the type 1 error}
}
\value{
A numeric vector with elements corresponding to lower limit, maximum
likelihood, and the upper limit for the proportion of ants coming from the
upper part of the nest.
}
\description{
Calculate confidence/credible interval for a vector of probabilities deriving
from the probability mass distribution
}
