% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colony_stats}
\alias{colony_stats}
\title{A data frame summarizing ant removal experiment across colonies}
\format{
Columns description:
\describe{
\item{N1}{number of workers from the upper part of the nest}
\item{N2}{number of workers from the lower part of the nest}
\item{marked_1}{number of marked workers from the upper part of the nest}
\item{marked_2}{number of marked workers from the lower part of the nest}
\item{removed_1}{number of removed workers from the uper part of the nest
which have been marked}
\item{removed_2}{number of removed workers from the lower part of the nest
which have been marked}
\item{removed total}{total number of removed workers}
\item{colony}{colony ID}
}
}
\usage{
colony_stats
}
\description{
A data frame summarizing ant removal experiment across colonies
}
\keyword{datasets}
