% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment_course}
\alias{experiment_course}
\title{A data frame presenting results of the ant collection event. Ants were
collected from the arena - they were active outside the nest.}
\format{
Columns description:
\describe{
\item{Date}{Collection date}
\item{Colony}{Colony ID}
\item{Unmarked}{Cumulative number of collected ants which have not
been marked}
\item{Red}{Cumulative number of the ants marked with red paint}
\item{Yellow}{Cumulative number of the ants marked with yellow paint}
\item{Dead_unm}{Cumulative number of the dead unmarked ants}
\item{Dead_red}{Cumulative number of the dead ants marked with red color}
\item{Dead_yellow}{Cumulative number of the dead ants marked with yellow
color}
\item{Removed_1}{Cumulative number of the removed marked ants from the upper
part of the nest}
\item{Removed_2}{Cumulative number of the removed marked ants from the lower
part of the nest}
\item{Upper_part_dead}{Cumulative number of the marked dead ants from the
upper part of the nest}
\item{Lower_part_dead}{Cumulative number of the marked dead ants from the
lower part of the nest}
}
}
\usage{
experiment_course
}
\description{
A data frame presenting results of the ant collection event. Ants were
collected from the arena - they were active outside the nest.
}
\keyword{datasets}
