% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{figure_6}
\alias{figure_6}
\title{Figure 6}
\usage{
figure_6(sample_size = 10000)
}
\arguments{
\item{sample_size}{A numeric indicating the number of samples to be taken from
the probability distribution of the proportion of ants from the upper nest segment}
}
\description{
Plot figure 6 from the publication. Note that the results of the Fig. 6 are
subject to some small random variation to avoid this effect you can set the
initial value of the generator seed.
}
