% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_experiment.R
\name{get_prob_mass}
\alias{get_prob_mass}
\title{Probability mass distribution}
\usage{
get_prob_mass(
  removed_1,
  removed_2,
  unmarked,
  marked_1,
  marked_2,
  N1,
  N2,
  ...,
  null_model = FALSE
)
}
\arguments{
\item{removed_1}{The number of marked and removed ants
which come from the upper part of the nest}

\item{removed_2}{The number of ants removed which come from the lower part of the nest}

\item{unmarked}{The number of removed ants which had not been marked}

\item{marked_1}{The number from the upper part of the nest which had been marked}

\item{marked_2}{The number from the lower part of the nest which had been marked}

\item{N1}{The total number of ants from the upper part of the nest}

\item{N2}{The total number of ants from the lower part of the nest}

\item{...}{Additional arguments to be omitted}

\item{null_model}{logical; if set to TRUE the probability mass distribution
for the null model is calculated; the null model assumes an equal probability of
being removed for an ant from the upper and lower part of the nest}
}
\value{
Vector of the length three accounting for the lower limit, minimum likelihood, and
the upper limit for the proportion of ants coming from the upper part of the nest
}
\description{
Calculate probability mass distribution for the number of individuals from
the one of two groups based on the hypergeometric distribution
}
\examples{
data(experiment_course)
data(ant_removal)
dataset <- dplyr::left_join(experiment_course,
select(ant_removal, colony, N1, N2, marked_1, marked_2))
do.call(get_prob_mass, as.list(unlist(dataset[2,])))
}
