% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main functions.R
\name{hgeom_numbers}
\alias{hgeom_numbers}
\title{Main}
\usage{
hgeom_numbers(
  alpha = 0.05,
  Unmarked,
  removed_1,
  removed_2,
  marked_1,
  marked_2,
  removed_total,
  N1,
  N2,
  ...
)
}
\arguments{
\item{alpha}{A number between 0 and 1 corresponding to the type 1 error}

\item{removed_1}{The number of ants removed which come from the upper part of the nest}

\item{removed_2}{The number of ants removed which come from the lower part of the nest}

\item{marked_1}{The number of ants marked and removed ants
which come from the upper part of the nest}

\item{marked_2}{The number of ants marked and removed ants
which come from the lower part of the nest}

\item{N1}{The total number of ants from the upper part of the nest}

\item{N2}{The total number of ants from the lower part of the nest}

\item{removal_total}{The total number of ants removed from the arena}
}
\value{
Vector of the length three accouting for the lower limit, maimum likelihood, and
the upper limit for the number of ants coming from the upper part of the nest
}
\description{
Caclulate confidence interval for the number  of ants from the upper part of the nest
taking into consideration all possible assignemnts of unmarked ants to either
upper or lower part group
}
\examples{
data(ant_removal)
do.call(hgeom_numbers, as.list(unlist(ant_removal[1,]))))
}
