% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{permutated_slopes}
\alias{permutated_slopes}
\title{Slopes obtained as a result of the permutation of the time series data}
\format{
An object of class \code{list} of length 5.
}
\usage{
permutated_slopes
}
\description{
Slopes obtained as a result of the permutation of the time series data
}
\keyword{datasets}
