% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_points}
\alias{reference_points}
\title{Reference points}
\format{
An object of class \code{data.frame} with 10 rows and 5 columns.
}
\usage{
reference_points
}
\description{
Data frame presenting distances between points used as reference to define the position of the
places (nest parts) where ants were found. Arrangement equal to one means that the order of the points
going clockwise was: A -> C -> B.
}
\keyword{datasets}
