% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_distribution}
\alias{spatial_distribution}
\title{Worker spatial distribution within the nest}
\format{
Columns description:
\describe{
\item{Colony}{Colony ID.}
\item{Point_ID}{ID of the place (nest part) where ants were collected.}
\item{Date}{Collection date. Note that different part of the colonies were
often collected in different days.}

\item{Distance to A}{Distance in cm from the central point of the excacated part of the nest
to the reference point A.}
\item{Distance to B}{Distance in cm from the central point of the excacated part of the nest
to the reference point B.}
\item{Distance to C}{Distance in cm from the central point of the excacated part of the nest
to the reference point C.}
\item{Radius}{The maximum distance from the central point of the excavated part of the nest
where ants were found and collected. Measured on horizontal plane.}
\item{Deprt from}{The minimum depth at which ants from the excavated part of the nest
were found and collected.}
\item{Deprt to}{The maximum depth at which ants from the excavated part of the nest
were found and collected.}
\item{Worker number}{Sum of the worker excavated from the corresponding part of the nest.}
}
}
\usage{
spatial_distribution
}
\description{
A data frame presenting results of the F. fusca colonies excavation during the winter.
}
\keyword{datasets}
