% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_analyses.R
\name{track_CI_change}
\alias{track_CI_change}
\title{Calculate credible interval}
\usage{
track_CI_change(time_series_result, colony_stats)
}
\arguments{
\item{time_series_result}{A dataset with the results of the ant removal experiment;
retrieved by \code{data("time_series_result", package = "overwintering")}.}

\item{colony_stats}{A dataset containing numeric parameters describing colony state and
experiment preparation;retrieved by \code{data("colony_stats", package = "overwintering")}.}
}
\value{
an updated version of of \code{time_series_results} data frame with CI values
added (long format)
}
\description{
Calculate credible interval for data obtained over the course of experiment
}
\details{
This function uses \code{time_series_results} and \code{colony_stats} datasets to
calculate how confidence interval for the null hypothesis and credible interval for
actual data changes over the course of experiment; this issue was examined following
the suggestion of the reviewer
}
