% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vert_distribution}
\alias{vert_distribution}
\title{A list of data frames representing data on different colonies.}
\format{
Each data frame contains following columns:
\describe{
\item{colony}{colony ID}
\item{depth}{depth range, in cm}
\item{worker_number}{worker number within given depth range}
\item{cumulative_number}{sum of worker number from actual and
upper depth ranges}
}
}
\usage{
vert_distribution
}
\description{
A list of data frames representing data on different colonies.
}
\keyword{datasets}
