% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{vertical_distribution_plot}
\alias{vertical_distribution_plot}
\title{Figure 3}
\usage{
vertical_distribution_plot(vert_distribution, colony_metadata)
}
\arguments{
\item{vert_distribution}{A list of data frames with the data on the density of ants along
vertical axis.}

\item{colony_metadata}{A list of lists with the various information about a colony. Retirved by
calling \code{data(colony_metadata)}}
}
\description{
Plot the main part of the Figure 3 from the publication.
}
\details{
Note that the total number of workers for part of colonies are greater that those
reported in the publication. The reason for this is that here we take into account all the
excavated workers whereas in the publication the sum was calculated from only
those individuals whose position was determined. As mentioned in the publication,
for some of the workers I failed to determine the depth at which they stayed before
excavation.
}
