% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main functions.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculate confidence interval for a vector of probabilities deriving from
the probability mass function}
\usage{
calculate_CI(data, alpha)
}
\arguments{
\item{data}{Vector of numbers representing probabilities; should sum up to 1}

\item{alpha}{Numeric value setting the mamximal value for the type 1 error}
}
\description{
Calculate confidence interval for a vector of probabilities deriving from
the probability mass function
}
