% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_experiment.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Vector of indices for the confidence interval}
\usage{
calculate_CI(data, alpha)
}
\arguments{
\item{data}{Data frame returned by \code{get_prob_mass}}

\item{alpha}{Numeric value setting the maximal value for the type 1 error}
}
\value{
Indices of the passed vector corresponding to the confidence interval
limits
}
\description{
Calculate confidence/credible interval for a vector of probabilities deriving
from the probability mass distribution
}
