% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_experiment.R
\name{get_prob_mass}
\alias{get_prob_mass}
\title{Probability mass distribution}
\usage{
get_prob_mass(
  removed_1,
  removed_2,
  unmarked,
  marked_1,
  marked_2,
  N1,
  N2,
  ...,
  null_model = FALSE
)
}
\arguments{
\item{removed_1}{The number of marked ants from the upper part of the nest which
were removed}

\item{removed_2}{The number of the marked ants from the lower part of the nest which
were removed}

\item{unmarked}{The number of the removed ants which had not been marked}

\item{marked_1}{The total number of the marked ants from the upper part of the nest}

\item{marked_2}{The total number of the marked ants from the lower part of the nest}

\item{N1}{The total number of ants from the upper part of the nest}

\item{N2}{The total number of ants from the lower part of the nest}

\item{...}{Additional arguments to be omitted}

\item{null_model}{logical; if set to TRUE the probability mass distribution
for the null model is calculated; the null model assumes an equal probability of
being removed for an ant from the upper and lower part of the nest; if set to FALSE
the probability takes into account the numbers of marked ans which have been removed
from the outside of nest and the model assumes that a marked ant has the same
probability of being removed as an unmarked one.}
}
\value{
A data frame with the all possible proportions of ants from the upper part
of the nest among all ants active outside the nest and the corresponding probabilities.
}
\description{
Calculate probability mass distribution for the proportion of ants from
the upper part of the nest among ants removed from outside the nest
}
\examples{
data(time_series_results)
data(colony_stats)
dataset <- dplyr::left_join(time_series_results,
dplyr::select(colony_stats, colony, N1, N2, marked_1, marked_2))
do.call(get_prob_mass, as.list(unlist(dataset[2,])))
}
